capstone/src/requirements.txt
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.11.1
backcall==0.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-bigquery==1.26.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.10
importlib-metadata==1.5.0
importlib-resources==1.3.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==6.1.2
jupyter-core==4.6.3
numpy==1.19.0
pandas==1.0.5
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pipx==0.15.1.3
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyspark==3.0.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.0
requests==2.24.0
rsa==4.6
six==1.14.0
soupsieve==2.0.1
tmdbv3api==1.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.9
userpath==1.3.0
virtualenv==20.0.10
wcwidth==0.1.9
zipp==3.1.0
